pyproject.toml
[project]
name = "notewiseai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
]
